Prevent quicksync double starts and cancels.

+++ src/qt/modaloverlay.cpp
@@ -55,6 +55,7 @@ ModalOverlay::ModalOverlay(const PlatformStyle *_platformStyle, QWidget *parent)
     showQuickSyncOptions = false;
 
 
+
     // DeepOnion: Theme
     ui->contentWidget->setStyleSheet(platformStyle->getThemeManager()->getCurrent()->getStyleSheet());
     ui->warningIcon->setIcon(platformStyle->SingleColorIcon(platformStyle->getThemeManager()->getCurrent()->getWarningIco()));
@@ -247,6 +248,9 @@ void ModalOverlay::refreshStyle()
 
 void ModalOverlay::onQuickSyncClicked()
 {
+    QString cstatus = getQuickSyncStatus();
+    if(cstatus != "" && cstatus != "Canceled")
+        return;
     setNetworkStatus(false);
     deflationrequested = true;
     quickSyncStatus = QuickSyncStatus::PREPARING;
@@ -263,6 +267,9 @@ void ModalOverlay::onQuickSyncClicked()
 
 void ModalOverlay::onCancelButtonClicked()
 {
+    QString cstring = getQuickSyncStatus();
+    if(cstring == "" || cstring == "Canceled")
+        return;
     deflationrequested = false;
     m_downloader.cancelDownload();
     fs::remove_all(tempquickSyncDir);
